requirements.txt
AfricastalkingGateway==1.8
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.33.0
django-cors-middleware==1.3.1
django-jet==1.0.6
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.0
gevent==1.1.1
greenlet==0.4.12
gunicorn==19.6.0
idna==2.6
infobip-api-python-client==2.0.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.3
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
UkPostcodeParser==1.1.1
uritemplate==3.0.0
urllib3==1.22
